refactor: compact app provider import path
